resolved issues with review
